+++ example/lib/features/image_gallery/domain/use_cases/clear_image_cache.dart
@@ -1,0 +1,12 @@
+import 'package:fpdart/fpdart.dart';
+
+import '../failures/gallery_failure.dart';
+import '../repositories/gallery_repository.dart';
+
+final class ClearImageCache {
+  const ClearImageCache(this._repository);
+
+  final GalleryRepository _repository;
+
+  Future<Either<GalleryFailure, Unit>> call() => _repository.clearCache();
+}